code review changes: - move logic to the auth init
