+++ lib/service/service.go
@@ -1770,18 +1770,6 @@ func (process *TeleportProcess) initAuthService() error {
 	var err error
 	cfg := process.Config
 
-	// We have to ensure that the process database is newly created before
-	// version update check.
-	var firstTimeStart bool
-	if _, err := process.storage.GetState(process.GracefulExitContext(), types.RoleAdmin); trace.IsNotFound(err) {
-		firstTimeStart = true
-	} else if err != nil {
-		return trace.Wrap(err)
-	}
-	if err := process.validateAndUpdateTeleportVersion(teleport.SemVersion, firstTimeStart); err != nil {
-		return trace.Wrap(err)
-	}
-
 	// Initialize the storage back-ends for keys, events and records
 	b, err := process.initAuthStorage()
 	if err != nil {
@@ -1924,6 +1912,7 @@ func (process *TeleportProcess) initAuthService() error {
 		process.ExitContext(),
 		auth.InitConfig{
 			Backend:                 b,
+			ProcessStorage:          process.storage,
 			Authority:               cfg.Keygen,
 			ClusterConfiguration:    cfg.ClusterConfiguration,
 			ClusterAuditConfig:      cfg.Auth.AuditConfig,
